refactor: fix utils-to-facades/services layer violations (closes #550)
